tcpc: Add driver for TI TUSB422 tcpc
Initial driver for TUSB422 TCPC which is a tcpci compliant TCPC. This
TCPC does not inlude a Type C mux and uses the tcpci driver for all of
the methods.

BUG=b:111281797
BRANCH=none
TEST=Verified operation as sink on DragonEgg. Have not verified source
operation, or low power/auto-toggle modes as those config options
can't be enabled on ITE at this point.

Change-Id: I783a5e2c4a13bc0b8fa4da4b134588382542024c
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1178994
Commit-Ready: Jett Rink <jettrink@chromium.org>
Tested-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>