usbc: hpd_deadline cleanup
some of the usages of this static variable were used in
overridable functions and they also had their own static
variable for this same purpose.  Since it is not required
to override all of the functions, this left two variables
out of sync with one another. So made them the same variable.

BUG=b:147535104
BRANCH=none
TEST=make buildall -j

Change-Id: Ic560d3a2a2e129450e918e0cb6dfff75fd1222a2
Signed-off-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2002953
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Edward Hill <ecgh@chromium.org>