PD: Add support for new host command, EC_CMD_GET_PD_PORT_CAPS
This host command is used to report the static capabilities of a USB-PD
port, including its power role, Try power role, data role, and its
physical port location on the device. This will be used to expose the
information in ACPI, via the EC object.

BUG=b:146506369
BRANCH=none
TEST=Along with coreboot changes, dump the SSDT, and verify
that the object looks as expected.

Change-Id: Ie1975a8b391eba6e924b0552ba9b0973fd2c63f3
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2015825
Reviewed-by: Denis Brockus <dbrockus@chromium.org>