zephyr: make IS_ENABLED compatible with shim
Since we are using Zephyr headers and EC sources together, we need to
make an IS_ENABLED macro which is compatible with both the Zephyr
concept of being enabled, and the CrOS EC concept (since we inherit
many default options from config.h).

The Zephyr concept of being enabled is defined, specifically only to
the token "1". The CrOS EC concept of being enabled is defined,
specifically to no tokens. Thus a macro which allows both concepts
defines "enabled" as defined, either to no tokens, or the token "1".

BUG=b:167590251
BRANCH=none
TEST=compile common/timer.c in follow-up CLs

Change-Id: Ieded05e65d0d7c27070f40ef1dd4d24db73b9de4
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2427093
Reviewed-by: Jett Rink <jettrink@chromium.org>