npcx9: workaround the download_from_flash API of the booter
The download_from_flash API in the booter is called to copy the RO/RW
image from flash to code RAM when doing the sysjump. There is a bug in
the npcx9 A1 chip. We have to workaround it with this CL and remove
the bypass when A2 chip is available.

BRANCH=none
BUG=b:165777478
TEST=pass "make buildall"
TEST="sysjump RO/RW" succeeds in the npcx9 EVB.

Signed-off-by: CHLin <CHLin56@nuvoton.com>
Change-Id: Id2babe9b9dbd36ca8b0450051d22632eb5bd4825
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2435165
Tested-by: CH Lin <chlin56@nuvoton.com>
Tested-by: caveh jalali <caveh@chromium.org>
Reviewed-by: caveh jalali <caveh@chromium.org>
Commit-Queue: CH Lin <chlin56@nuvoton.com>