TCPMv2: Create charge allow_list and add devices
The charge allow_list is for partner devices that
do not advertise Unconstrained Power in their
SRC_Caps but we should still charge from them
if they are the SRC.

The allow_list should be constrained to partners
that perform dual role and do not present
Unconstrained Power in their SRC_Caps. It is
preferred for the partner to send this
correctly but for products that are already out
in the wild, it may be a really bad user
experience for this to fail.

BUG=b:170644500,b:170848849,b:170970249
BRANCH=zork
TEST=connect E24d monitor and verify it charges DUT

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Ib4f2c62af3bb2c1ed066ea4353ec26cd536a4c26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2490729
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>